requirements.txt
versioned-hdf5==1.6.0
iopath==0.1.10
matplotlib==3.7.1
nibabel==5.2.0
numpy==1.24.3
pandas==1.5.3
Pillow==9.4.0
pytest==7.4.0
pytorch_msssim==1.0.0
pytz==2022.7
PyYAML==6.0
scipy==1.10.1
seaborn==0.13.0
setuptools==68.0.0
SimpleITK==2.3.0
simplejson==3.19.1
scikit-image==0.20.0
tabulate==0.8.10
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
visdom==0.2.4
